commonJS style
