+++ lib/index.js
@@ -2,4 +2,6 @@
 
 require('setimmediate')
 
-window.storj = require('./core');
+const Storj = require('./core');
+
+exports = module.exports = Storj
